refactor: Exclude "search" and "index" pages from generateStaticParams
This commit updates the generateStaticParams function in page.jsx to exclude pages with the slugs "search" and "index". Previously, all pages were included in the generated static params, but this caused issues with the rendering of certain blocks. By excluding these specific pages, we ensure that only the necessary pages are included in the static params.